avxjudge: Speed up instruction lookup by using sets instead of lists
Python's `in` operator are translated to a hash table lookup in Python;
it's more efficient than using it with lists or tuples.  Modify the
instruction tables to use sets instead.

Also changes how the instruction tables are initialized; initializing
them using this syntax is slighly more efficient.

Analyzing the libc library on my system went from 3.40s to 2.61s with
this change alone.

Signed-off-by: Leandro Pereira <leandro.pereira@intel.com>