clr-avx-move.pl: Avoid renaming what we've already renamed
If a package is built with both AVX512 and AVX2, when the
%make_install_avx2 target is run, we'll already have AVX512 files in
/usr/bin/haswell, /usr/lib64/haswell and plugins named *.so.avx512. So
don't recurse into the libsubdir and don't rename anything called
*.so.avx*.

This patch is a little bigger than necessary so we don't recurse at all
for binary directories.

Note: this only works if AVX512 is run first, because it depends on the
AVX512 directory being a subdir of the AVX2 directory.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>