Fix moving when there are symlinks to symlinks
That's the usual case with libraries:
  libfoo.so -> libfoo.so.1 -> libfoo.so.1.2.3

We prepended the symlinks so we did readelf on them first, before
moving. But there was a problem in case moved the middle symlink, such
that the first one became broken.

So instead of building a file list to scan, then scan and move in one
go, we scan while building the file list, doing all the moves at the
end.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>