elf-move: Fix --path behavior and crash
If --path was not passed previously elf-move would crash, it also was
incorrectly handling the filepath instead of the virtpath (only care
about the path after the DESTDIR for --path, makes handling easier
too).

Finally make --path override the --skip behavior so they can work
together to append new content to a file.