Skip setuid files
The setuid files need to be specially annotated in the spec
file in order for the setuid attribute to be handled correctly by
rpm. Because the optimized binaries are not easily able to be
identified this way in the spec file, just skip setuid binaries from
being added to the optimized list.

Signed-off-by: William Douglas <william.douglas@intel.com>