Directly install lib types
The libraries do not need to conditionally installed as they are used
conditionally already. While this takes up additional disk space it is
preferrable for now due to swupd not handling the missing files well.

This requires a change in how targetpath is interpretted. Now the
targetpath is no longer the path given to the optimized file directory
but instead the path to the base of the target and the optimzied file
directory is now fixed ('/usr/share/clear/optimized-elf').

Signed-off-by: William Douglas <william.douglas@intel.com>