Rework elf-move to move content into a path prefix
The packages now are expected to have content placed in a optimized
path prefix, V3 for AVX2 and V4 for AVX512 (with additional prefixes
as new options are added). Update elf-move.py to move content in the
installdir to the targetdir under a prefix based on the btype.

Library files that are moved to the hwcaps directory are not impacted
by this change. All other files that are moved will now be found in
the targetdir with the remaining fullpath name of prefix + virtpath.

Signed-off-by: William Douglas <william.douglas@intel.com>